Fix: InputChat.tsx 스타일 서식 모자란 부분 수정 #4
