+++ src/components/ChatHeader.tsx
@@ -1,8 +1,7 @@
 import { IconLeft } from "../IconData";
-
 const ChatHeader = () => {
     return (
-        <div className="flex p-5 w-full">
+        <div className="select-none flex p-5 w-full border border-gray-600">
             <button type="button" className="text-zero fill-white w-5">
                 <IconLeft />
                 뒤로가기
